sio: use a ring buffer to do asynchronous IO
Warning. The buffer is a fixed size. If it fills, we panic. The only
other option I know of (and perhaps the more sane option) is to block,
but I think it is better to prove that we never actually have to, since
we are not supposed to use such loops in out project.